+++ lib/views/pages/main_tab_controller_page.dart
@@ -4,9 +4,9 @@ import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:imperio/stores/sports_store.dart';
 import 'package:imperio/utils/service_locator.dart';
 import 'package:imperio/views/pages/all_sports_page.dart';
-import 'package:imperio/views/widgets/gradient_background.dart';
-import 'package:imperio/views/widgets/upper_tab_bar.dart';
-import 'package:imperio/views/widgets/upper_tab_button.dart';
+import 'package:imperio/views/widgets/shared/gradient_background.dart';
+import 'package:imperio/views/widgets/upper_tab/upper_tab_bar.dart';
+import 'package:imperio/views/widgets/upper_tab/upper_tab_button.dart';
 
 class MainTabController extends StatefulWidget {
   const MainTabController({super.key});
