+++ lib/views/widgets/sport_banner/sports_banner.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:imperio/models/featured_banner.dart';
-import 'package:imperio/views/widgets/sport_banner_card.dart';
+import 'package:imperio/views/widgets/sport_banner/sport_banner_card.dart';
 
 class SportsBanner extends StatelessWidget {
   final List<FeaturedBanner> banners = [
